pyproject.toml
[tool.poetry]
name = "quipubase"
version = "0.0.1"
description = "A Document Database based on `json_schema` on top of RocksDB with VectorStore features."
authors = ["bahamondex <obahamondemunoz@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
orjson = "^3.10.16"
rocksdict = "^0.3.25"
pydantic = "^2.11.3"
fastapi = "^0.115.12"
faiss-cpu = "^1.10.0"
sse-starlette = "^2.3.3"
light-embed = "^1.0.7"
uvicorn = {extras = ["standard"], version = "^0.34.2"}
aioredis = "^2.0.1"
python-dotenv = "^1.1.0"
anydocs = {git = "https://github.com/bahamondex/anydocs.git"}
cbase64 = "^0.0.9"
python-multipart = "^0.0.20"
anthropic = "^0.52.0"
openai = "^1.82.0"
boto3 = "^1.38.23"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
boto3-stubs = {extras = ["s3"], version = "^1.38.23"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"